requirements-posterior.txt
tqdm
numpy
hydra-core
jax
chex
equinox
openml
blackjax
ucimlrepo
